refactor: one route file for checklistItem
